Refactor Player into small components and more Volume icons
Adding two new icons to Volumen, low and medium
